sokol_app.h: fix undefined behaviour (shift on signed value, fixes #616)

+++ sokol_app.h
@@ -6609,7 +6609,7 @@ _SOKOL_PRIVATE LRESULT CALLBACK _sapp_win32_wndproc(HWND hWnd, UINT uMsg, WPARAM
                     POINT point;
                     GetCursorPos(&point);
                     ScreenToClient(_sapp.win32.hwnd, &point);
-                    PostMessage(_sapp.win32.hwnd, WM_MOUSEMOVE, 0, point.x|(point.y << 16));
+                    PostMessage(_sapp.win32.hwnd, WM_MOUSEMOVE, 0, (uint32_t)point.x)|(((uint32_t)point.y) << 16));
                 }
                 break;
             case WM_DROPFILES:
